+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.3.7
+
+* Add links on the RPC to the artist, track and album
+
+## 1.3.6
+
+* Upgraded packages
+* Icon should now be shown on Linux
+* Add support for customizable status name display (https://github.com/CuteTenshii/deezer-discord-rpc/pull/123)
+* Fixed some CSS classes
+
 ## 1.3.5
 
 * Fixed the RPC (well, this time it should be fixed [I hope])